requirements.txt
pydantic>=2.7,<3.0
PyYAML>=6.0,<7.0
rapidfuzz>=3.9,<4.0
